Move beatmap play button nearer its only remaining usage
