+++ osu.Game/Overlays/BeatmapSet/Buttons/PreviewButton.cs
@@ -11,7 +11,6 @@ using osu.Game.Audio;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osu.Game.Online.API.Requests.Responses;
-using osu.Game.Overlays.BeatmapListing.Panels;
 using osuTK;
 
 namespace osu.Game.Overlays.BeatmapSet.Buttons
